requirements.txt
matplotlib
numpy
scipy
pyserial
tk
